Match minified/unminified bootstrap versions
The minified and un-minified Bootstrap versions tracked in the repo
differ: the minified version is `v3.x`, while the un-minified version is
`v5.x-beta`.

This caused a breakage in layout when switching to the minified
versions. See #273.

Bring both the JS and CSS minified versions into parity with the
unmified versions.

Also see #267