Label Neovim logo for accessibility
Add both `aria-label` property and the SVG `<title>` tag to mark up the
graphic for screen reader accessibility. Use both for the widest
support.

Add `role="img"` to the SVG, since SVGs do not have graphics semantics
by default, and the logo is used as a graphic (obviously), so marking it
as one gets us a lot of proper accessibility for free.

Add the exact width and height in dedicated properties instead of inline
`style`; the `width` and `height` properties are static, so the browser
can calculate the size of the image before layout rendering, increasing
performance.

Re-tab the navigation HTML file, bringing the logo call into the same
spaces-as-indentation style as the rest of the file.