Prevent flash of invisible text
There is a period during page load where the web font is not downloaded,
so text is invisible. If there is a network hiccup, this state could
be permanent for the session.

Modern browsers support `font-display`, which controls strategies for
how to handle this problem.

Use the `swap` property, which strikes a balance between preferring the
specified web font, but quickly falling back to a default if it does not
load in time, so text will always be visible.

Older browsers will ignore the property, maintaining the current
behavior.

See https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/font-display