feat: enable BlueSky account verification
This commit adds a 'https://neovim.io/.well-known/atproto-did' page with
the content of a DID value taken from dedicated Neovim BlueSky account.
With this page present, account will be able to use a `@neovim.io`
handle which assures users that this is an "official" BlueSky account of
Neovim team.

Details:
- The '.well-known' is a long established dedicated directory name for
  automated verification that certain domain belongs to a certain
  account on third-party site.
- The DID value of an account is a public information, so there should
  not be issues with it being present here in the open.