feat: add rss link to document head for autodiscovery
In order to not end up with a double slash in the path when
constructing the absolute feed url out of `.Site.BaseURL` and
`.Site.Params.news.feed`, I've inserted a dot in-between. In a http
request, double slashes will be transmitted as-is, potentially causing
issues, whereas dot segments are to be resolved by the client beforehand
(as per rfc3986). It's the little things… ;-))